+++ src/colmap/controllers/hierarchical_mapper.h
@@ -29,10 +29,10 @@
 
 #pragma once
 
-#include "colmap/controllers/base_controller.h"
 #include "colmap/controllers/incremental_mapper.h"
 #include "colmap/scene/reconstruction_manager.h"
 #include "colmap/scene/scene_clustering.h"
+#include "colmap/util/base_controller.h"
 
 #include <memory>
