+++ src/colmap/controllers/incremental_mapper.h
@@ -29,9 +29,9 @@
 
 #pragma once
 
-#include "colmap/controllers/base_controller.h"
 #include "colmap/scene/reconstruction_manager.h"
 #include "colmap/sfm/incremental_mapper.h"
+#include "colmap/util/base_controller.h"
 
 namespace colmap {
