+++ src/colmap/mvs/fusion.h
@@ -29,7 +29,6 @@
 
 #pragma once
 
-#include "colmap/controllers/base_controller.h"
 #include "colmap/math/math.h"
 #include "colmap/mvs/depth_map.h"
 #include "colmap/mvs/image.h"
@@ -37,6 +36,7 @@
 #include "colmap/mvs/model.h"
 #include "colmap/mvs/normal_map.h"
 #include "colmap/mvs/workspace.h"
+#include "colmap/util/base_controller.h"
 #include "colmap/util/cache.h"
 #include "colmap/util/eigen_alignment.h"
 #include "colmap/util/ply.h"
