+++ src/colmap/mvs/patch_match.h
@@ -34,7 +34,7 @@
 #include "colmap/mvs/model.h"
 #include "colmap/mvs/normal_map.h"
 #ifndef __CUDACC__
-#include "colmap/controllers/base_controller.h"
+#include "colmap/util/base_controller.h"
 #include "colmap/util/threading.h"
 #endif
