+++ src/colmap/util/CMakeLists.txt
@@ -33,6 +33,7 @@ set(FOLDER_NAME "util")
 COLMAP_ADD_LIBRARY(
     NAME colmap_util
     SRCS
+        base_controller.h base_controller.cc
         cache.h
         controller_thread.h
         eigen_alignment.h
