+++ src/colmap/util/base_controller.h
